client_pimpl::send_auth_response: fix warning

+++ src/wsclient.cpp
@@ -331,7 +331,7 @@ namespace ws {
 		auto auth = b64decode(auth_msg);
 
 		if (!auth_msg.empty()) {
-			inbufs[0].cbBuffer = auth.size();
+			inbufs[0].cbBuffer = (unsigned long)auth.size();
 			inbufs[0].BufferType = SECBUFFER_TOKEN;
 			inbufs[0].pvBuffer = auth.data();
